jamf-health-tool/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jamf-health-tool"
version = "0.1.0"
description = "CLI for Jamf Pro health checks: policy failures, profile scope audits, and MDM command failures."
authors = [{ name = "Generated by ChatGPT Codex" }]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "tabulate>=0.9",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]
reports = [
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
]

[project.scripts]
jamf-health-tool = "jamf_health_tool.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["jamf_health_tool"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
